movign pramas to fsdp config
