+++ src/axolotl/utils/schemas/config.py
@@ -33,6 +33,7 @@ from axolotl.utils.schemas.integrations import (
     WandbConfig,
 )
 from axolotl.utils.schemas.internal import EnvCapabilities, GPUCapabilities
+from axolotl.utils.schemas.fsdp import FSDPConfig
 from axolotl.utils.schemas.model import (
     ModelInputConfig,
     ModelOutputConfig,
@@ -367,13 +368,6 @@ class AxolotlInputConfig(
             "used in combination with torch.compile."
         },
     )
-    fp8_enable_fsdp_float8_all_gather: bool | None = Field(
-        default=None,
-        json_schema_extra={
-            "description": "Enable FSDP float8 all-gather optimization for FP8 training. Can "
-            "improve training speed by 10-15% when FSDP is enabled."
-        },
-    )
     bfloat16: bool | None = Field(
         default=None,
         json_schema_extra={
@@ -656,24 +650,9 @@ class AxolotlInputConfig(
             "description": "Whether to use deepcompile for faster training with deepspeed"
         },
     )
-    fsdp: list[str] | None = Field(
-        default=None,
-        json_schema_extra={"description": "FSDP configuration"},
-        deprecated="Configuring FSDP using `fsdp` is deprecated. Please use `fsdp_config` instead. ",
-    )
-    # TODO @SalmanMohammadi strongly type this as its own schema
-    fsdp_config: dict[str, Any] | None = Field(
+
+    fsdp_config: FSDPConfig | None = Field(
         default=None, json_schema_extra={"description": "FSDP configuration options"}
-    )
-    fsdp_version: int | None = Field(
-        default=None,
-        json_schema_extra={"description": "FSDP version"},
-    )
-    fsdp_final_state_dict_type: (
-        Literal["FULL_STATE_DICT", "LOCAL_STATE_DICT", "SHARDED_STATE_DICT"] | None
-    ) = Field(
-        default=None,
-        deprecated="Configuring FSDP final state dict type using `fsdp_final_state_dict_type` is deprecated. Please use `fsdp_config.final_state_dict_type` instead.",
     )
 
     val_set_size: float | None = Field(
@@ -1272,23 +1251,6 @@ class AxolotlConfigWCapabilities(AxolotlInputConfig):
 
         if version.parse(torch_version) < version.parse("2.6.0"):
             raise ValueError("QAT is not supported on torch version < 2.6.0")
-
-        return data
-
-    @model_validator(mode="before")
-    @classmethod
-    def check_fsdp_torch_version(cls, data):
-        env_capabilities = data.get("env_capabilities", {})
-        torch_version = env_capabilities.get("torch_version")
-
-        if torch_version is None:
-            import torch
-
-            torch_version = str(torch.__version__).split("+", maxsplit=1)[0]
-
-        if data.get("fsdp_config") and str(data.get("fsdp_version")) == "2":
-            if version.parse(torch_version) < version.parse("2.7.0"):
-                raise ValueError("FSDP2 is not supported on torch version < 2.7.0")
 
         return data
